requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
yt-dlp>=2023.11.16
pinecone>=5.0.0
sentence-transformers>=2.2.2
pydantic>=2.5.0
aiohttp>=3.9.1
numpy>=1.24.3
python-multipart==0.0.6
requests>=2.31.0
torch>=2.0.0
transformers>=4.35.0
rank-bm25>=0.2.2

